infra/src/cloud_function/00_load_file_to_bq/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
google-api-core==2.*
ndjson==0.3.1
jq ==1.8.0